Revert "Merge branch 'master' of https://github.com/azerothcore/azerothcore-wotlk"
This reverts commit ae1955ec80f6ddc5c533f4bd33210150b2b0d862, reversing
changes made to e46f225807df8ea049f405701d96496182676655.

+++ src/server/game/Handlers/QueryHandler.cpp
@@ -393,7 +393,7 @@ void WorldSession::HandlePageTextQueryOpcode(WorldPacket& recvData)
                     ObjectMgr::GetLocaleString(player->Text, loc_idx, Text);
 
             data << Text;
-            data << pageText->NextPage;
+            data << uint32(pageText->NextPage);
             pageID = pageText->NextPage;
         }
         SendPacket(&data);
